feat: implement responsive sidebar component with collapsible functionality
- Added a new SidebarComponent to manage navigation and sidebar display.
- Integrated SidebarComponent into various views (light, mobile, sim, super, side).
- Updated styles for the sidebar to ensure responsiveness and proper display on mobile devices.
- Modified existing components to utilize the new sidebar structure, enhancing the overall layout.
- Added conditional rendering for buttons based on user login status in create-web component.
- Refactored HTML structure in multiple components for better semantic organization.
- Improved accessibility and user experience with better handling of sidebar visibility and interactions.
