+++ package.json
@@ -23,18 +23,18 @@
   "type": "module",
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.2.11",
-    "@angular/common": "^19.2.11",
-    "@angular/compiler": "^19.2.11",
-    "@angular/core": "^19.2.11",
-    "@angular/forms": "^19.2.11",
-    "@angular/platform-browser": "^19.2.11",
-    "@angular/platform-browser-dynamic": "^19.2.11",
-    "@angular/router": "^19.2.11",
-    "@angular/service-worker": "^19.2.11",
+    "@angular/animations": "^20.1.0",
+    "@angular/common": "^20.1.0",
+    "@angular/compiler": "^20.1.0",
+    "@angular/core": "^20.1.0",
+    "@angular/forms": "^20.1.0",
+    "@angular/platform-browser": "^20.1.0",
+    "@angular/platform-browser-dynamic": "^20.1.0",
+    "@angular/router": "^20.1.0",
+    "@angular/service-worker": "^20.1.0",
     "axios": "^1.10.0",
     "chrome-finder": "^1.0.7",
-    "cron": "^4.3.1",
+    "cron": "^4.3.2",
     "dayjs": "^1.11.13",
     "express": "^5.1.0",
     "file-saver": "^2.0.5",
@@ -43,20 +43,20 @@
     "localforage": "^1.10.0",
     "lz-string": "^1.5.0",
     "mitt": "^3.0.1",
-    "ng-zorro-antd": "^19.3.1",
+    "ng-zorro-antd": "^20.0.0",
     "nprogress": "^0.2.0",
-    "puppeteer": "^24.10.2",
+    "puppeteer": "^24.12.1",
     "qs": "^6.14.0",
     "rough-notation": "^0.5.1",
     "rxjs": "~7.8.2",
-    "sharp": "^0.34.2",
+    "sharp": "^0.34.3",
     "zone.js": "~0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.2.12",
-    "@angular/cli": "^19.2.12",
-    "@angular/compiler-cli": "^19.2.11",
-    "@tailwindcss/postcss": "^4.1.10",
+    "@angular-devkit/build-angular": "^20.1.0",
+    "@angular/cli": "^20.1.0",
+    "@angular/compiler-cli": "^20.1.0",
+    "@tailwindcss/postcss": "^4.1.11",
     "@types/compression": "^1.8.1",
     "@types/file-saver": "^2.0.7",
     "@types/jasmine": "~5.1.8",
@@ -71,12 +71,12 @@
     "info-web": "^0.0.46",
     "jasmine-core": "~5.8.0",
     "lint-staged": "^16.1.2",
-    "nodemailer": "^7.0.3",
-    "oxlint": "^1.3.0",
+    "nodemailer": "^7.0.5",
+    "oxlint": "^1.6.0",
     "pm2": "^6.0.8",
     "postcss": "^8.5.6",
-    "prettier": "^3.6.0",
-    "tailwindcss": "^4.1.10",
+    "prettier": "^3.6.2",
+    "tailwindcss": "^4.1.11",
     "tslib": "^2.8.1",
     "tsx": "^4.20.3",
     "typescript": "~5.8.3"
