+++ src/view/side/index.component.scss
@@ -3,23 +3,15 @@
     padding-right: 12px !important;
     padding-left: 10px !important;
   }
-  .sider-compnent {
-    position: fixed !important;
-    top: 0 !important;
-    left: 0 !important;
-    height: 100% !important;
-    transform: translateX(-100%);
-  }
 }
 
-.layout {
+.main {
   min-height: 100vh;
-  &.showSidebar {
-    .sider-compnent {
-      transform: translateX(0) !important;
-    }
-  }
+  background-color: #f0f2f5;
+  display: flex;
   .inner-layout {
+    flex: 1;
+    width: 0;
     .content {
       margin-bottom: 2rem;
       padding: 10px 70px 0 15px;
@@ -28,53 +20,11 @@
     }
   }
 
-  .mask {
-    z-index: 31;
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: 100vh;
-    backdrop-filter: blur(3px);
-  }
-
   .box {
     flex: 1;
     transition: 0.1s linear;
   }
-
-  .collapse-wrapper {
-    position: absolute;
-    right: 0;
-    top: 0;
-    padding: 0 15px;
-
-    .collapseIcon {
-      display: block;
-      transform: rotate(-270deg);
-
-      &.active {
-        transform: rotate(-360deg);
-      }
-    }
-  }
-
-  .ant-menu-item {
-    padding-right: 40px;
-    padding-left: 38px !important;
-  }
   ::ng-deep {
-    .ant-menu.ant-menu-inline-collapsed {
-      & > .ant-menu-submenu {
-        > .ant-menu-submenu-title {
-          padding: 0 !important;
-          text-align: center;
-        }
-      }
-      .sideicon {
-        margin-right: 0;
-      }
-    }
     .web-list {
       padding: 0;
       .name {
@@ -84,80 +34,5 @@
         border-radius: 2px !important;
       }
     }
-  }
-}
-
-.logo {
-  z-index: 11;
-  width: 100%;
-  background-color: #fff;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  cursor: pointer;
-  user-select: none;
-  text-align: center;
-  line-height: 70px;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.05);
-  img {
-    max-height: 50px;
-  }
-  .logo-img {
-    width: 50px;
-    height: 50px;
-    pointer-events: none;
-  }
-  .web-title {
-    font-weight: bold;
-    font-size: 20px;
-    margin-left: 8px;
-    vertical-align: middle;
-  }
-}
-
-.sider-compnent {
-  z-index: 32;
-  position: sticky;
-  top: 0;
-  left: 0;
-  height: 100vh;
-  background-color: #fff;
-  width: 220px;
-  ::ng-deep {
-    .ant-layout-sider-children {
-      display: flex;
-      flex-direction: column;
-    }
-  }
-  .mobile-collapse {
-    position: absolute;
-    top: 50px;
-    right: -40px;
-    width: 40px;
-    height: 40px;
-    box-shadow: 2px 0 8px #00000026;
-    background-color: #fff;
-    border-radius: 0 4px 4px 0;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-  }
-}
-
-.sider {
-  height: 100%;
-  overflow: hidden;
-  overflow-y: auto;
-  flex: 1;
-  padding-bottom: 30px;
-  .menu-title {
-    max-width: 130px;
-  }
-}
-
-::ng-deep {
-  .ant-menu-title-content {
-    display: inline-flex;
-    align-items: center;
   }
 }
